requirements.txt
streamlit
groq
SpeechRecognition
gTTS
pydub
simpleaudio
google.generativeai
simpleaudio